Change type of size_ field of LocalOp from long to size_t

+++ itensor/mps/localop.h
@@ -37,7 +37,7 @@ class LocalOp
     Tensor const* Op2_;
     Tensor const* L_;
     Tensor const* R_;
-    mutable long size_;
+    mutable size_t size_;
     public:
 
     using IndexT = typename Tensor::index_type;
@@ -350,7 +350,7 @@ long inline LocalOp<Tensor>::
 size() const
     {
     if(!(*this)) Error("LocalOp is default constructed");
-    if(size_ == -1)
+    if(size_ == size_t(-1))
         {
         //Calculate linear size of this 
         //op as a square matrix
